+++ WalletWasabi.Fluent/ViewModels/NavBar/INavBarButton.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+
+namespace WalletWasabi.Fluent;
+
+public interface INavBarButton : INavBarItem
+{
+	Task Activate();
+}
