+++ WalletWasabi.Fluent/ViewModels/NavBar/INavBarToggle.cs
@@ -1,0 +1,6 @@
+﻿namespace WalletWasabi.Fluent;
+
+public interface INavBarToggle : INavBarItem
+{
+	void Toggle();
+}
