+++ WalletWasabi.Fluent/ViewModels/NavBar/NavBarPosition.cs
@@ -1,0 +1,8 @@
+﻿namespace WalletWasabi.Fluent;
+
+public enum NavBarPosition
+{
+	None,
+	Top,
+	Bottom
+}
