+++ WalletWasabi.Fluent/ViewModels/NavBar/NavBarSelectionMode.cs
@@ -1,0 +1,8 @@
+﻿namespace WalletWasabi.Fluent;
+
+public enum NavBarSelectionMode
+{
+	Selected = 0,
+	Button = 1,
+	Toggle = 2
+}
