+++ WalletWasabi.Fluent/ViewModels/NavBar/NavigationTarget.cs
@@ -1,0 +1,10 @@
+﻿namespace WalletWasabi.Fluent;
+
+public enum NavigationTarget
+{
+	Default = 0,
+	HomeScreen = 1,
+	DialogScreen = 2,
+	FullScreen = 3,
+	CompactDialogScreen = 4,
+}
