+++ WalletWasabi.Fluent/ViewModels/Settings/PrivacyModeViewModel.cs
@@ -1,11 +1,13 @@
 using System.Reactive.Linq;
 using ReactiveUI;
+using WalletWasabi.Fluent.ViewModels.Navigation;
 
 namespace WalletWasabi.Fluent.ViewModels.Settings;
 
-[NavigationMetaData(Title = "Discreet Mode", Searchable = false, NavBarPosition = NavBarPosition.Bottom, NavBarSelectionMode = NavBarSelectionMode.Toggle)]
-public partial class PrivacyModeViewModel : ViewModelBase
+[NavigationMetaData(Searchable = false, NavBarPosition = NavBarPosition.Bottom, NavBarSelectionMode = NavBarSelectionMode.Toggle)]
+public partial class PrivacyModeViewModel : RoutableViewModel
 {
+	private string _title = "";
 	[AutoNotify] private bool _privacyMode;
 	[AutoNotify] private string? _iconName;
 	[AutoNotify] private string? _iconNameFocused;
@@ -14,15 +16,24 @@ public partial class PrivacyModeViewModel : ViewModelBase
 	{
 		_privacyMode = Services.UiConfig.PrivacyMode;
 
+		SetTitleAndIcon();
+
 		this.WhenAnyValue(x => x.PrivacyMode)
 			.Skip(1)
 			.ObserveOn(RxApp.TaskpoolScheduler)
 			.Subscribe(x => Services.UiConfig.PrivacyMode = x);
 	}
 
+	public override string Title
+	{
+		get => _title;
+		protected set => this.RaiseAndSetIfChanged(ref _title, value);
+	}
+
 	public void Toggle()
 	{
 		PrivacyMode = !PrivacyMode;
+		SetTitleAndIcon();
 	}
 
 	public void SetTitleAndIcon()
